Bug fix: failed.sql getting truncated if there are some failed DDLs from post snapshot import phase

+++ yb-voyager/cmd/importSchema.go
@@ -208,7 +208,7 @@ func isYBDatabaseIsColocated(conn *pgx.Conn) bool {
 
 func dumpStatements(stmts []string, filePath string) {
 	if len(stmts) == 0 {
-		if utils.FileOrFolderExists(filePath) {
+		if !flagPostSnapshotImport && utils.FileOrFolderExists(filePath)==true {
 			err := os.Remove(filePath)
 			if err != nil {
 				utils.ErrExit("remove file: %v", err)
@@ -218,7 +218,13 @@ func dumpStatements(stmts []string, filePath string) {
 		return
 	}
 
-	file, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644)
+	var fileMode int
+	if flagPostSnapshotImport {
+		fileMode = os.O_WRONLY | os.O_CREATE | os.O_APPEND
+	} else {
+		fileMode = os.O_WRONLY | os.O_CREATE | os.O_TRUNC
+	}
+	file, err := os.OpenFile(filePath, fileMode, 0644)
 	if err != nil {
 		utils.ErrExit("open file: %v", err)
 	}
